+++ src/main/java/com/kcv/account/management/repository/IPackageRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPackageRepository extends JpaRepository<PackageDTO, Integer>{
+public interface IPackageRepository extends JpaRepository<PackageDTO, Long>{
 
 	
 }
